requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
albumentations==1.4.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
av==12.1.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
boilerpy3==1.0.7
boto3==1.34.115
botocore==1.34.115
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
cattrs==23.2.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
cleo==2.1.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
construct==2.5.3
contourpy==1.2.0
crashtest==0.4.1
cryptography==42.0.7
cycler==0.10.0
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
deep-sort-realtime==1.3.2
deepsort==0.0.5
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
dulwich==0.21.7
email_validator==2.1.1
et-xmlfile==1.1.0
Events==0.5
executing==2.0.1
farm-haystack==1.25.5
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.3
filterpy==1.4.5
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.5.4
gcloud==0.18.3
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-cloud-texttospeech==2.16.3
google-cloud-translate==3.15.3
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gpt4all==2.7.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
gTTS==2.5.1
h11==0.14.0
h5py==3.10.0
haystack==0.42
haystack-ai==2.1.2
haystack-bm25==1.0.2
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==2.10
ifaddr==0.2.0
imageio==2.34.0
imgviz==1.7.5
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.2.1
install==1.3.5
installer==0.7.0
ipykernel==6.29.4
ipython==8.22.2
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.15.0
keyring==24.3.1
kiwisolver==1.4.5
kubernetes==29.0.0
labelme==5.4.1
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-google-vertexai==1.0.4
langchain-text-splitters==0.2.0
langsmith==0.1.67
lazy-imports==0.3.1
lazy_loader==0.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mlxtend==0.23.1
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
norfair==2.2.0
num2words==0.5.13
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.30.5
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
opt-einsum==3.3.0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.3
pefile==2023.2.7
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.11.0
platformdirs==4.2.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
posthog==3.5.0
prompt-toolkit==3.0.43
prompthub-py==4.0.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.6
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.3
pydeck==0.9.1
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
python-ptrace==0.9.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
QtPy==2.4.1
quantulum3==0.9.1
rank-bm25==0.2.2
rapidfuzz==3.9.3
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==12.6.0
roboflow==1.1.30
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
sahi==0.11.16
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.37.2
streamlit==1.35.0
streamlit-webrtc==0.47.7
supervision==0.21.0
sympy==1.12
tenacity==8.3.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
trove-classifiers==2024.5.22
typeguard==4.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.25
url-normalize==1.4.3
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vertexai==1.49.0
virtualenv==20.26.2
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
xattr==1.1.0
yarl==1.9.4
yolov5==7.0.13
yolov8==0.0.2
zipp==3.19.1
